requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
langgraph==0.2.60
openai==1.58.1
pdfplumber==0.11.4
python-docx==1.1.2
sqlalchemy==2.0.36
pymysql==1.1.1
python-dotenv==1.0.1